🐛 Prevent GPUs from raising OOM when loading checkpoints for distributed training
The GPUs were crashing because each process' GPU is loading the optimizer's state dict to "cuda:0", the GPU where the state dict was originally saved at. This leads to "cuda:0" being overloaded and OOMs on each process, and explains the behaviour where this failure mode is only observed with num_gpus > 2 during distributed training.